fix ci
fix ci

changed onehot in tensor_compute

fix ci

+++ src/kernels/cpu/reference/onehot.cpp
@@ -40,7 +40,7 @@ result<void> onehot_impl(const int32_t *indices, T *output, const runtime_shape_
             indices_index[i] = out_index[i + 1];
         }
         auto index = indices[offset(get_default_strides(indices_shape), indices_index)];
-        T out_v = index == out_index[axis] ? on_value : off_value;
+        T out_v = static_cast<size_t>(index) == out_index[axis] ? on_value : off_value;
         output[offset(out_strides, out_index)] = out_v;
         return ok();
     });
